finish off baritone integration
